refactor(shared,wallet): make shared foundational, break shared→accounts cycle
shared/cashu.ts and shared/spark.ts imported up into accounts. Move getSeedPhraseDerivationPath (a pure BIP-85 path helper) from accounts/account-cryptography.ts to shared/cryptography.ts and delete the now-empty file. Move useTrackAndUpdateSparkAccountBalances (a React reactivity hook reading account-hooks; only consumer was wallet.tsx) from shared/spark.ts to wallet/use-track-spark-account-balances.ts. shared/spark.ts is now pure Spark-wallet utilities with no accounts import. The only remaining domain-layer cycle is shared↔agicash-db, to be resolved during the SDK move.

Pure relocation, no behavior change.

Part of the wallet-SDK no-cache migration (cycle-break before move).
