+++ apps/web-wallet/app/features/shared/cashu.ts
@@ -37,8 +37,8 @@ import { useMemo } from 'react';
 import { getQueryClient } from '~/features/shared/query-client';
 import { measureOperation } from '~/lib/performance';
 import { computeSHA256 } from '~/lib/sha256';
-import { getSeedPhraseDerivationPath } from '../accounts/account-cryptography';
 import { getAgicashMintAuthProvider } from './agicash-mint-auth-provider';
+import { getSeedPhraseDerivationPath } from './cryptography';
 
 // Cashu-specific derivation path with hardnened indexes to derive public keys for
 // locking mint quotes and proofs. 129372 is UTF-8 for 🥜 (see NUT-13) and the other
