+++ apps/web-wallet/app/features/wallet/use-track-spark-account-balances.ts
@@ -1,0 +1,57 @@
+import type { SdkEvent } from '@agicash/breez-sdk-spark';
+import { Money } from '@agicash/money';
+import { useEffect } from 'react';
+import { useAccounts, useAccountsCache } from '../accounts/account-hooks';
+import { sparkDebugLog } from '../shared/spark';
+
+export function useTrackAndUpdateSparkAccountBalances() {
+  const { data: sparkOnlineAccounts } = useAccounts({
+    type: 'spark',
+    isOnline: true,
+  });
+  const accountCache = useAccountsCache();
+
+  useEffect(() => {
+    const registrations = sparkOnlineAccounts.map((account) => {
+      const listenerPromise = account.wallet.addEventListener({
+        onEvent(event: SdkEvent) {
+          sparkDebugLog('Breez event', {
+            accountId: account.id,
+            type: event.type,
+          });
+
+          if (
+            event.type === 'paymentSucceeded' ||
+            event.type === 'paymentPending' ||
+            event.type === 'paymentFailed' ||
+            event.type === 'claimedDeposits' ||
+            event.type === 'synced'
+          ) {
+            account.wallet.getInfo({}).then((info) => {
+              const balance = new Money({
+                amount: info.balanceSats,
+                currency: 'BTC',
+                unit: 'sat',
+              }) as Money;
+              accountCache.updateSparkAccountBalance({
+                accountId: account.id,
+                balance,
+              });
+            });
+          }
+        },
+      });
+      return { wallet: account.wallet, listenerPromise };
+    });
+
+    return () => {
+      for (const { wallet, listenerPromise } of registrations) {
+        listenerPromise
+          .then((id) => wallet.removeEventListener(id))
+          .catch(() => {
+            console.warn('Failed to remove Spark event listener');
+          });
+      }
+    };
+  }, [sparkOnlineAccounts, accountCache]);
+}
